# contract/openapi.yaml
openapi: 3.1.0
info:
  title: Identity Service
  version: 2.0.0
  description: >
    Central identity and access management service.
    Handles member registration, authentication (email/password + social OAuth),
    legacy member migration, session management, token verification,
    organisational hierarchy, group management, and permissions.
servers:
  - url: http://localhost:3000
    description: Local development
components:
  schemas:
    ErrorResponse:
      type: object
      properties:
        error:
          type: object
          properties:
            code:      { type: string }
            message:   { type: string }
            requestId: { type: string }
            details:
              type: array
              items:
                type: object
                properties:
                  field:   { type: string }
                  message: { type: string }
                required: [message]
          required: [code, message, requestId, details]
      required: [error]
...
    Installation:
      type: object
      properties:
        id:         { type: string }
        name:       { type: string }
        code:       { type: string }
        timezone:   { type: string }
        is_active:  { type: boolean }
        created_at: { type: string, format: date-time }
      required: [id, name, code, timezone, is_active, created_at]
    Nation:
      type: object
      properties:
        id:                { type: string }
        installation_id:   { type: string }
        installation_name: { type: string, nullable: true }
        installation_code: { type: string, nullable: true }
        name:              { type: string }
        description:       { type: string, nullable: true }
        is_active:         { type: boolean }
        created_at:        { type: string, format: date-time }
      required: [id, installation_id, name, is_active, created_at]
    Tribe:
      type: object
      properties:
        id:                { type: string }
        nation_id:         { type: string }
        nation_name:       { type: string, nullable: true }
        installation_name: { type: string, nullable: true }
        installation_code: { type: string, nullable: true }
        name:              { type: string }
        description:       { type: string, nullable: true }
        is_active:         { type: boolean }
        created_at:        { type: string, format: date-time }
      required: [id, nation_id, name, is_active, created_at]
    WorkforceTeam:
      type: object
      properties:
        id:          { type: string }
        name:        { type: string }
        description: { type: string, nullable: true }
        is_active:   { type: boolean }
        created_at:  { type: string, format: date-time }
      required: [id, name, is_active, created_at]
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: >
        JWT issued by the identity service.
        Obtain via /v1/auth/register, /v1/auth/login, or OAuth flows.
paths:
...
  /v1/org/installations:
    post:
      summary: Create installation
      operationId: createInstallation
      tags: [Organisation]
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, code]
              properties:
                name:     { type: string, minLength: 1 }
                code:     { type: string, minLength: 1 }
                timezone: { type: string, default: UTC }
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                type: object
                properties:
                  data: { $ref: '#/components/schemas/Installation' }
        '400': { description: Validation error, content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } } }
        '401': { description: Unauthorized, content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } } }
        '409': { description: Code exists, content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } } }
    get:
      summary: List installations
      operationId: listInstallations
      tags: [Organisation]
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          description: List
          content:
            application/json:
              schema:
                type: object
                properties:
                  data: { type: array, items: { $ref: '#/components/schemas/Installation' } }
        '401': { description: Unauthorized, content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } } }
  /v1/org/installations/{id}:
    get:
      summary: Get installation by ID
      operationId: getInstallation
      tags: [Organisation]
      security: [{ bearerAuth: [] }]
      parameters: [{ name: id, in: path, required: true, schema: { type: string } }]
      responses:
        '200':
          description: Installation
          content:
            application/json:
              schema:
                type: object
                properties:
                  data: { $ref: '#/components/schemas/Installation' }
        '401': { description: Unauthorized, content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } } }
        '404': { description: Not found, content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } } }
    delete:
      summary: Deactivate installation
      operationId: deleteInstallation
      tags: [Organisation]
      security: [{ bearerAuth: [] }]
      parameters: [{ name: id, in: path, required: true, schema: { type: string } }]
      responses:
        '204': { description: Deactivated }
        '401': { description: Unauthorized, content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } } }
        '404': { description: Not found, content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } } }
  /v1/org/nations:
    post:
      summary: Create nation
      operationId: createNation
      tags: [Organisation]
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, installationId]
              properties:
                name:           { type: string, minLength: 1 }
                installationId: { type: string }
                description:    { type: string }
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                type: object
                properties:
                  data: { $ref: '#/components/schemas/Nation' }
        '400': { description: Validation error, content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } } }
        '401': { description: Unauthorized, content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } } }
    get:
      summary: List nations
      operationId: listNations
      tags: [Organisation]
      security: [{ bearerAuth: [] }]
      parameters: [{ name: installationId, in: query, required: false, schema: { type: string } }]
      responses:
        '200':
          description: List
          content:
            application/json:
              schema:
                type: object
                properties:
                  data: { type: array, items: { $ref: '#/components/schemas/Nation' } }
        '401': { description: Unauthorized, content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } } }
  /v1/org/nations/{id}:
    get:
      summary: Get nation by ID
      operationId: getNation
      tags: [Organisation]
      security: [{ bearerAuth: [] }]
      parameters: [{ name: id, in: path, required: true, schema: { type: string } }]
      responses:
        '200':
          description: Nation with installation context
          content:
            application/json:
              schema:
                type: object
                properties:
                  data: { $ref: '#/components/schemas/Nation' }
        '401': { description: Unauthorized, content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } } }
        '404': { description: Not found, content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } } }
    delete:
      summary: Deactivate nation
      operationId: deleteNation
      tags: [Organisation]
      security: [{ bearerAuth: [] }]
      parameters: [{ name: id, in: path, required: true, schema: { type: string } }]
      responses:
        '204': { description: Deactivated }
        '401': { description: Unauthorized, content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } } }
        '404': { description: Not found, content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } } }
  /v1/org/tribes:
    post:
      summary: Create tribe
      operationId: createTribe
      tags: [Organisation]
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, nationId]
              properties:
                name:        { type: string, minLength: 1 }
                nationId:    { type: string }
                description: { type: string }
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                type: object
                properties:
                  data: { $ref: '#/components/schemas/Tribe' }
        '400': { description: Validation error, content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } } }
        '401': { description: Unauthorized, content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } } }
    get:
      summary: List tribes
      operationId: listTribes
      tags: [Organisation]
      security: [{ bearerAuth: [] }]
      parameters: [{ name: nationId, in: query, required: false, schema: { type: string } }]
      responses:
        '200':
          description: List
          content:
            application/json:
              schema:
                type: object
                properties:
                  data: { type: array, items: { $ref: '#/components/schemas/Tribe' } }
        '401': { description: Unauthorized, content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } } }
  /v1/org/tribes/{id}:
    get:
      summary: Get tribe by ID
      operationId: getTribe
      tags: [Organisation]
      security: [{ bearerAuth: [] }]
      parameters: [{ name: id, in: path, required: true, schema: { type: string } }]
      responses:
        '200':
          description: Tribe with nation and installation context
          content:
            application/json:
              schema:
                type: object
                properties:
                  data: { $ref: '#/components/schemas/Tribe' }
        '401': { description: Unauthorized, content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } } }
        '404': { description: Not found, content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } } }
    delete:
      summary: Deactivate tribe
      operationId: deleteTribe
      tags: [Organisation]
      security: [{ bearerAuth: [] }]
      parameters: [{ name: id, in: path, required: true, schema: { type: string } }]
      responses:
        '204': { description: Deactivated }
        '401': { description: Unauthorized, content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } } }
        '404': { description: Not found, content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } } }
  /v1/org/workforce-teams:
    post:
      summary: Create workforce team
      operationId: createWorkforceTeam
      tags: [Organisation]
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name:        { type: string, minLength: 1 }
                description: { type: string }
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                type: object
                properties:
                  data: { $ref: '#/components/schemas/WorkforceTeam' }
        '400': { description: Validation error, content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } } }
        '401': { description: Unauthorized, content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } } }
    get:
      summary: List workforce teams
      operationId: listWorkforceTeams
      tags: [Organisation]
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          description: List
          content:
            application/json:
              schema:
                type: object
                properties:
                  data: { type: array, items: { $ref: '#/components/schemas/WorkforceTeam' } }
        '401': { description: Unauthorized, content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } } }
  /v1/org/workforce-teams/{id}:
    get:
      summary: Get workforce team by ID
      operationId: getWorkforceTeam
      tags: [Organisation]
      security: [{ bearerAuth: [] }]
      parameters: [{ name: id, in: path, required: true, schema: { type: string } }]
      responses:
        '200':
          description: Workforce team
          content:
            application/json:
              schema:
                type: object
                properties:
                  data: { $ref: '#/components/schemas/WorkforceTeam' }
        '401': { description: Unauthorized, content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } } }
        '404': { description: Not found, content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } } }
    delete:
      summary: Deactivate workforce team
      operationId: deleteWorkforceTeam
      tags: [Organisation]
      security: [{ bearerAuth: [] }]
      parameters: [{ name: id, in: path, required: true, schema: { type: string } }]
      responses:
        '204': { description: Deactivated }
        '401': { description: Unauthorized, content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } } }
        '404': { description: Not found, content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } } }
...